requirements.txt
anyio==3.7.1
asttokens==2.4.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distlib==0.3.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
executing==2.0.1
fastapi==0.104.1
filelock==3.13.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
icecream==2.1.3
identify==2.5.31
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jose==1.0.0
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
platformdirs==3.11.0
pluggy==1.3.0
pre-commit==3.5.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.16.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
ruff==0.1.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
sqlalchemy2-stubs==0.0.2a36
sqlmodel==0.0.11
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
virtualenv==20.24.6
